Grid fix to regrid_sphere2
Changed the grid to similar definition as in regrid_sphere.  The CESM
definition of 64 x 128 for T42 grids is such to prevent aliasing of
non-linear terms.  For the truncation only about 43 latitude grid points
are necessary (ntrunc + 1) in this case.  Grid points are now defined by
supplied ntrunc, and altered to remain even.