LMR_utils global_mean update
Updated the global mean function in LMR_utils.  Right now changes are
located in global_mean2 until I figure out all the dependencies of the
original function.  Main change is that it can take either a simple latitude
vector, a lat/lon shaped grid of latitude, or a flattened grid of latitude.
Must match the spatial dimensions of the given field or method fails.