Calibration on AnalysisVariable
Fixed the calibration of our linear psm to work with AnalysisVariable
class.  It will load when it cannot find calibration in a pre_calibration
dataset.  Preliminary testing makes the current calibration seem slightly
different than what we've been using in our pre_calibrated data file.
I'm not sure if this is a big problem yet.