PSM Calibration with full proxy record
I updated the Proxy initialization to keep the full proxy record until
after it has created the object so if calibration is required, it can
do so against whatever years the proxy and calibration have in common.
As far as I can tell this code replicates the old calibration exactly.
Only differences in calibration must come from differences in the calibratio
file, but I checked against the old load_gridded_methods and it loads
them verbatim according to my test.  For now I think the linear psm values
are close enough to the pre_calibrated files I have that I will move forward.