Calibration obj passing and gridded float32 fix
Made some minor bug fixes to passing Calibration objects when loading
all proxies.  Also fixed a calibration error due to lat/lon being
stored as float32 instead of float64.  For now, assume that lat's are
exact and deserve as much numerical precision as possible.  There are
still differences between this and old calibration style due to how
old proxies were loaded.  For now it does not appear to be a huge issue
for the DA process.