Variance Adjustment
Implemented necessary features for forecast variance adjustment testing.
Two methods are employed currently.  The first is a prior covariance
hybrid between the forecast and the static initial prior.  The other is
adjustment of the forecast mapping G's eigenmodes to nudge some of the
modes into exponential growth instead of decay.