Fixed Numpy Backend, State list type change
I specified the np.roll axis and this fixed the reconstruction differences
between this and the h5 backend.  Updated the state to no longer be a list of
ndarrays, and it's just an np array.  Needed to do this because it was changing
types throughout its existence.  Makes copying the state easier.