Seed and multi_seed added to config.
Added wrapper parameter class for adding knobs to the LMR_wrapper.
Right now it houses multi_seed which can be used to set a
pre-determined list of seeds equal in length to the number of
realizations.  Seed is added to the core configuration.  It's
overridden by the multi_seed if running using the wrapper.  The seed
value is propagated to the proxies and prior class for use in the
random sampling.