Precalculated Ye Loading into Driver
Altered the driver to now try and load pre-calculated yes for the
offline case.  If it doesn't find the file or the flag to use this
feature is False, then it falls back to the old method.

Added a precalc file loading function to utils and also created a
new error that I can use to escape from the try block if the flag
says to not to use the pre calculated file.

Config added a new parameter to the core class called 'use_precalc_ye'