Fix lat/lon index finding bug in pre_building ye
I was improperly using the np.unravel_index shape causing the
grid point finding code to give the wrong index.  Right now it's
set as a very non general case where it expects the ordering
of dimensions to be lat then lon, and that these are the trailing
dimensions.

Wrote tests to make sure the grid point finder, and subsequent
ye calculations and ye file match the original method used in
the driver.