Fix build ye proxy id to index map
The index map was improperly using the proxy id output of the load
all function for ProxyPages.  The id output is grouped by proxy type
while we want the proxy id's in order of the actual proxy objects.
Just iterate through the proxy objects to retrieve ids and order now.