Fix Multi-seed and curr_iter
The multi-seed set would previously only change the core.seed.
Because this value is propagated to the prior and the proxies upon
instantiation of a Config() class I needed to change these.

Now for each iteration the wrapper instantiates a new configuration
with the correct seeds.  This fixed discrepancies between multi-seed
reconstructions of 500hpa heights between pre-calc and non-pre-calc
ye values.