Add in ensemble-mean forecast flag
Added a configuration flag that allows for user to try a LIM forecast
on only the ensemble mean.  The ensemble mean is removed from each
individual member, and then after forecasting is re-added to individual
members.  The idea here was to try and preserve some aspects of the
ensemble variance.  Experimental feature right now.