Parameter Search Integration
Took the parameter search features used in the experimental
branches and meshed it into a form more consistent with how the
LMR framework is built.

Added a wrapper class into the configuration which uses multi-seed
and also param_search.  Multi-seed is for consistent but random
monte-Carlo iterations.  Param_search will perform sub iterations
for each combination of parameters in an MC iteration.

Added some util functions for setting the configuration parameters
to LMR_utils2.  Handles taking the parameter string (e.g. core.hybrid_a)
and translating that to changing the actual parameter in the cfg object.