Fix temperature sensitivty map
Fixed the temperature sensitivity mapping in LMR_utils for loading
precalculated ye values.  There was an extra colon in 'temperature'
that I removed.  Also updated the text in the failure case where
a state variable is required for the psm but there is no precalc file
and it's not in the state vector.  Text now emphasizes that you should
just be able to run the precalculating script.