Formatting Fixes
Started fixing some of the formatting for the updates to Bilinear
and PDSI additions, but I think I'm just going to leave it for now.

When I say formating, I have it set for ~80 character limit in my
IDE.  I try to keep it in these bounds, but it's getting more
difficult for some of the long embedded statements.