Handling Non-defined Proxy Measurement/Type map
Changed the ProxyPages object to handle the case where the
proxy type/measurement mapping is undefined from the configuration.
There are a few records in the dataframe that are never used that
have odd measurement definitions.  These may eventually be removed,
but currently I changed to just leave the proxy_type as None.

Also changed verbosity to 1 in the driver.