Legacy config option and src path
Updated the configuration file with two general fields for
the code.  LEGACY_CONFIG provides the option to opt out of using
the yaml configuration, and just uses the LMR_config.py file.
SRC_DIR is the absolute path to the source code.  This helps loading
of the YAML files, and eventually will be used by all the tests.