FIX: pole inclusion for regridding funcs
The previous regridding style would always include the poles when
regridding using spherical harmonics.  This was incorrect according
to the documentation on github (jswhit/pyshparm).  If the number
of latitudes is even, equator and poles are not included, and
vice versa.  Both simple and spherical regridding now reflect this.
ESMpy regridding will not include the poles in its grids.