Cyclic point and irregular spacing
Added detection of the cyclic longitude not occuring at the edges
of the array.  This was breaking the monotonic increase detection.
It now rotates the data and longitude grids so that the cyclic point
occurs at the end of the longitude dimension for ESMpy regridding.

Dropped requirement that data be regularly spaced for input data
into the regridding operation.  The TOS data from ccsm4 is on a
rotated pole which does not have regularly spaced latitude values.
Not sure if the edge bounds will be completely correct, but there's
no support for output of grids using this function.