Regridded masked values
Updated to create mask on the destination grid after regridding.
Masked values are not updated on the destination grid, but are left
as 0.0.  If performing a regrid on masked input data, any zeros on
the destination field are changed to NaNs, and then the final output
is converted to a numpy masked array.