Python3 PSM_build and proxy ingest fixes
Handling of ascii files updated to be compatible with python 3.
Adjusted to read in NCDC templated files with UTF-8 encoding, and
to replace any un-encodable chars to '?'.  Pickle files now dump
to file objects specified as binary dumps.  Adjustment to integer
division for the global hemispheric means function in LMR_utils.