Fix ESMpy latitude shift for irregular grids
This is a fix for regridding irregularly gridded fields. Prior
to this if there was a shift in longitudes for the cyclic point, the
latitudes would not be shifted. This lead to a discrepency in the
rotated pole grids used for CCSM4 ocean fields. This adds a line
that now shifts the lat grid along with the lon grid.