Use lazy loading and fix eof spatial field
Loading in the entire netCDF used too much memory.  Instead leaves
the data unloaded and uses the for loop going over iterations and
ensemble members to load that chunk and calculate indices.  I believe
this was how it originally computed the values.  Also a small fix for
how the EOF was loaded back into an uncompressed (NaN for masked
gridpoints) data structure.