build_ye logging config & cleanup raised errors
In build_ye.py, there is now a if at the bottom to catch when code is
run from terminal. This should prevent errors from the script grabbing
unintended system arguments if some poor soul tries to import it. This
also has a very basic config setup for the logging module. It now
prints logging messages to terminal that were previously ignored.

There were a few cases where an "error" message (might be) printed and
then SystemExit. I changed these to throw a KeyError(message) so that we
can get a full stack trace. There was a case where the code simply quits
without a message and it can easily be mistaken for successful
completion of the script.

In LMR_proxy.py, the logged ERROR message was change to a print to be
more consistent with how errors are handled elsewhere in that module.