Major update to foram d18O PSMs
Options for the d18O PSMs can be set in `bayesreg_d18o` class, which are
described in the class docstrings. These can also be set in a YAML
config file. Key points:

* You can set `psm_d18osw_variable` to a salinity or seawater
d18O field in the state vector. If it's salinity,
`psm_d18osw_from_salinity` must be `True` so that the PSM can predict
seawater d18O from salinity. If we plug d18Osw in directly from the
state vector than use `False` -- this will be much faster to run.

* There is also a `seasonal_seatemp` config option. Setting this to
`True` tells the PSM to use the calibration based on seasonal SSTs and
not annual-means. `False` is just an annual-mean calibration which could
be useful for Deep Time DAs.

A d18O PSM was added for *N. incompta* (`BayesRegD18oIncomptaPSM` or
`bayesreg_d18o_incompta`). So we can use species-specific calibration
for these samples now. Cool stuff.

A d18O PSM was added for the "pooled" calibration. This is
`bayesreg_d18o_pooled` or `BayesRegD18oPooledPSM`. This is a catch-all
calibration that might be useful for non-exant species -- say, for a Deep
Time reanalysis.

This commit also changes the way the pre-calculated Ye files are named,
reflecting the wide range of options available for these PSMs.

There we several bug fixes to how seasonality was handled in d18O PSMs.