Can now give optional prior_timescale config in addn' to recon_timescale.
This lets users vary the timescale for binning the prior while using
a different timescale for binning the proxies and reconstructed
posterior. For example, we can now reconstruct with a timescale of 200
year bins, using a prior ensemble from 50-year average bins. You can set
with `cfg.prior.prior_timescale`. It must be > 1 or 'None' (!!null in
yaml). If `prior.prior_timescale` is not given or 'None', then it uses the
value for `core.recon_timescale` -- replicating old behavior.